refactor(web): move app actions into seperate directory
