+++ apps/web/src/app/layout.tsx
@@ -3,8 +3,8 @@ import type { Metadata } from 'next'
 import { Inter, Ubuntu_Mono } from 'next/font/google'
 import { clsx } from 'clsx'
 import { PublicEnvScript } from 'next-runtime-env'
-import { getLeftPanelPreferenceAction } from '@/components/layout/app-shell/app-shell-actions'
-import { getDarkModePreferenceAction } from '@/utils/dark-mode.actions'
+import { getLeftPanelPreferenceAction } from '@/actions/left-panel'
+import { getDarkModePreferenceAction } from '@/actions/dark-mode'
 import { onUnhandledRequest } from '@/__mocks__/utils/on-unhandled-request'
 import '@/styles/tailwind.css'
 import '@/styles/globals.scss'
